got-ocr-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
transformers==4.44.0
torch==2.4.0
torchvision==0.19.0
Pillow==10.4.0
PyMuPDF==1.24.0
tiktoken==0.7.0
